src/main/java/leetCode/RotateArray.java
//https://leetcode.com/problems/rotate-array/ (Right Rotation)
//https://www.hackerrank.com/challenges/ctci-array-left-rotation/problem (Left Rotation)

package leetCode;

public class RotateArray {
    public int[] rotateRight(int[] nums, int k) {
        k = k % nums.length;
        reverse(nums, 0, nums.length - 1);
        reverse(nums, 0, k - 1);
        reverse(nums, k, nums.length - 1);
        return nums;
    }

    public int[] rotateLeft(int[] nums, int k) {
        k = k % nums.length;
        reverse(nums, 0, nums.length - 1);
        reverse(nums, 0, nums.length - k - 1);
        reverse(nums, nums.length - k, nums.length - 1);
        return nums;
    }

    private void reverse(int[] nums, int start, int end) {
        while (start < end) {
            int temp = nums[start];
            nums[start] = nums[end];
            nums[end] = temp;
            start++;
            end--;
        }
    }
}

